wizard: Fix test that is broken on Jenkins.
This test mysteriously did not fail locally, but failed on Jenkins 
because an Application instance was created which caused a different
code path of execution.
This change uses AndroidTestBase as the test base class instead of 
TestCase. An Application instance is created to mimic the results seen 
on Jenkins.
The bug fix is to test for isUnitTestMode() and NOT require to be 
running on UI thread.

Change-Id: Ie06f3414b5abbdfcec37df8f2ced0182a6e1d8d9
(cherry picked from commit 47974026b953395640d8e0a389152b663db8cc42)
(cherry picked from commit 89c991e857e0641943252ede82c700a4f2e32d67)