Fixed bug where inter-module dependencies where not configured correctly.
This CL fixes a regression introduced by
https://android-review.googlesource.com/145656 . In that CL the IDE
sets a dependency on the 'exploded AAR' instead of setting a dependency
on a module, only if the module dependency does not have an AndroidProject
model.

It turns out that that check for AndroidProject was done too soon, when
modules where not fully configured, and even if a module had an
AndroidProject, the checked failed, because it was done before such
module can be configured with an AndroidProject.

The fix is to set up dependency on 'exploded AAR' after all the modules
have been fully configured.

Fixes issue
https://code.google.com/p/android/issues/detail?id=163888

Change-Id: I1224659f3f6e31eea4940e7a3df034edfbfed2e5
(cherry picked from commit f520c921f24c53e1cbc0e040d71961d25bb455fb)