Now we check if cached project models are missing, if they are, we trigger a sync.
The issue is that if serialization/deserialization of a model fails, IDEA
swallows the exception and leaves the project in inconsistent state. This CL
verifies that the cache has all the needed models, and if a model is missing,
a project sync will be triggered.

Change-Id: Ia9e01f1e57abc6f30b8f4db1b4d10362893c1acb
(cherry picked from commit 94a375efc6ee4c8a06d8f3ef5a146ddd6a6e381f)