Fixed serialization of Java library models.
In Android Studio 1.2.0, it appears that project
syncing never ends, or at least the IDE is not being
notified when a sync ends. This left the
"Sync" button disabled, preventing users from
doing any work.

This is indeed a strange error, and it is caused
by several issues:

1. A classpath issue in the distribution of Android
Studio, but it does not show up when running
Studio from IDEA, and our automated UI tests were
not able to catch this issue.

2. This classpath problem is a Gradle class that was not
found when deserializing the Gradle model (a nice
feature introduced in IDEA 14.1 that can speed up
reopening projects.) In addition to the class not
being found in the Studio distribution, the caching
mechanism did not complain, but instead ignored
the exception.

3. Studio has a mechanism to check that the cache
is in valid state, so when #2 happens we can invoke
a Gradle sync, but the problem is that the IDE
still thought we were obtaining the model from the
cache, and it never enabled the "Sync" button
after a sync.

This CL fixes issues #1, by having the model for
Java libraries contain only JVM and IDEA classes
(e.g. Strings and java.io.Files.) This way we won't
ever experience a ClassNotFoundException when
deserializing the model.

Issue #2 is not that critical, and needs to be
addressed in IDEA itself.

A separate CL will fix issue #3.

Fix for
https://code.google.com/p/android/issues/detail?id=160899

Change-Id: I036f75d8cfcbaf8afbb1e448133721902734548a
(cherry picked from commit 4ea698c54b76a484509e62e63b9bbfea518afa23)