Fixed how project sync handles invalid cached models.
When Studio fails to deserialize a Gradle model stored in
disk, the cached model may be missing information. When
this happens, we detect that and trigger a full project
sync. The problem is, that the IDE still thinks the
model comes from the cache, and not from Gradle and
the "Sync" button is never enabled, and the "Sync in
progress" editor notification never goes away, giving
users the impression that project sync never ends.

This is the second part for fixing
https://code.google.com/p/android/issues/detail?id=160899

The first part of the fix is
https://android-review.googlesource.com/149093

Change-Id: I76dc461543d8ee2fd599b69e4dbd04297551c9ef
(cherry picked from commit 7c9f8f2c996b2cdb54c3ba204731cea52f87f2de)