Add the right locale to RenderParams.
In case of RTL locales, the locale was always set as "ur" and this
caused localisation bugs in widgets like CalendarView.

Also remove the incorrectly added line params.setRtlSupport(). The
method is supposed to indicate that the application supports RTL by
specifying the right flag in the manifest. For apps that don't have the
attribute set, we shouldn't lie about how the app will look in the
preview.

Change-Id: Icd85ce80e056126c7e3e91582258a2afe700e177