Not asserting when upgrading previews
When adding an update package to an UpdatablePkgInfo we weren't
checking for the case where the local package is a preview and the
remote isn't. Switched to simply using the base install ids, since
the extra checks were unnecessary (it would never be the case that
we'd have clashing base install ids for non-matching packages).
Fixes https://code.google.com/p/android/issues/detail?id=183068

Change-Id: I2a167727d0b74159476c4a48af076b2ea0d83103